pyproject.toml
[tool.poetry]
name = "example-bot"
version = "0.1.0"
description = ""
authors = ["Ilia Kaziamov <kaziamov@outlook.com>"]
readme = "README.md"
packages = [{include = "example_bot"}]

[tool.poetry.dependencies]
python = "^3.8"
aiogram = "^2.25.1"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
